refactor(deploy): split R2 helpers into focused modules with tests
The internal r2.ts had drifted into a grab bag of unrelated concerns
(SigV4 client, manifest naming, content-type table, XML escape). This
splits the file along its natural seams so each piece is testable on
its own and the R2 client stays focused on the wire protocol.

- xml.ts: predefined-entity escape/unescape. New xml.spec.ts covers
  round-trip + order-of-replacement (the &amp; pass must run last on
  decode).
- manifest.ts: release manifest key naming + the IMMUTABLE_PREFIX /
  RELEASE_MANIFEST_PREFIX constants. manifest.spec.ts moved over from
  r2.spec.ts.
- content-type.ts: extension-to-MIME table. content-type.spec.ts moved
  over and picks up a no-extension fallback case + a case-insensitive
  case.
- r2.ts: now only the R2Client + r2FromEnv. Imports the xml helpers
  from xml.ts. Renamed the inline byte iterator (`b` → `byte`) flagged
  by DeepSource.
- Renamed `d` → `parsed` in manifestKeyToDate and `t` → `when` in the
  prune retention specs to clear the remaining JS-C1002 findings.
- Updated import paths in prune-immutable-r2.ts, prune-immutable-r2.spec.ts,
  and sync-immutable-to-r2.ts.
